User Logout Finished

+++ resources/item.py
@@ -1,5 +1,6 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import create_access_token, get_jwt, get_jwt_identity, jwt_required
+import jwt
 from models.item import ItemModel
 
 class Item(Resource):
@@ -22,7 +23,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not found'}, 404
 
-    @jwt_required()
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -38,7 +39,14 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        print(">>>>>>")
+        print(claims)
+        if not claims['is_admin']:
+            return { "message" : "Only admin can delete the items."}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -61,6 +69,15 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            "item" : [item['name'] for item in items],
+            "message" : "Please log in for more information"
+        }, 200
+        
         # return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
-        return {'items': [x.json() for x in ItemModel.find_all()]}
